fix(fix errors): fix boundary error with dismiss client and server id in TextInput component
fix boundary error with dismiss client and server id in TextInput component | fix deprecated strict
mode   error
