+++ src/components/main/ui/Input.tsx
@@ -7,7 +7,7 @@ import React, {
   ChangeEvent,
 } from 'react';
 
-import { nanoid } from 'nanoid';
+// import { nanoid } from 'nanoid';
 import ArrowIcon from '@/components/icons/ArrowIcon';
 
 interface TextInputProps {
@@ -25,7 +25,7 @@ const TextInput: FC<TextInputProps> = ({
   options = [],
   ...rest
 }) => {
-  const id = nanoid();
+  // const id = nanoid();
   const [inputValue, setInputValue] = useState('');
   const [isOpen, setIsOpen] = useState(false);
   const dropdownRef = useRef<HTMLDivElement>(null);
@@ -66,7 +66,7 @@ const TextInput: FC<TextInputProps> = ({
       className={`w-full ${errorText ? 'text-red-500' : 'text-inherit'}`}
     >
       {title && (
-        <label htmlFor={id} className="text-sm font-medium">
+        <label className="text-sm font-medium">
           {title}
         </label>
       )}
@@ -79,7 +79,6 @@ const TextInput: FC<TextInputProps> = ({
         />
         <input
           {...rest}
-          id={id}
           value={inputValue}
           data-category={category}
           className={`h-[64px] w-full p-2 pl-12 placeholder:text-xl focus:outline-none ${
